+++ src/components/FormItems/Radio.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallowMatchSnapshot } from '../../common/test_utils';
-import Radio from './Radio';
+import Radio from './Radio.component';
 
 describe('Radio tests', () => {
   test('default renderer', () => {
