+++ src/components/FormItems/TextField.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallowMatchSnapshot } from '../../common/test_utils';
-import TextField from './TextField';
+import TextField from './TextField.component';
 
 describe('TextField tests', () => {
   test('default renderer', () => {
